#2363 Split out AuthorizationPostInitializer
This creates an initializer to be run separately, if needed, after all
servlets have been discovered. Splitting this up allows us to create the
authorization service early, but not yet commit the permissions.
