+++ extension/security-jakarta/src/main/java/cloud/piranha/extension/security/jakarta/JakartaSecurityAllInitializer.java
@@ -27,9 +27,17 @@
  */
 package cloud.piranha.extension.security.jakarta;
 
+import static cloud.piranha.extension.exousia.AuthorizationPreInitializer.AUTHZ_FACTORY_CLASS;
+import static cloud.piranha.extension.exousia.AuthorizationPreInitializer.AUTHZ_POLICY_CLASS;
+
 import java.util.Set;
 
-import cloud.piranha.extension.security.servlet.ServletSecurityAllInitializer;
+import org.glassfish.exousia.modules.def.DefaultPolicy;
+import org.glassfish.exousia.modules.def.DefaultPolicyConfigurationFactory;
+
+import cloud.piranha.extension.eleos.AuthenticationInitializer;
+import cloud.piranha.extension.exousia.AuthorizationInitializer;
+import cloud.piranha.extension.exousia.AuthorizationPreInitializer;
 import cloud.piranha.extension.soteria.SoteriaInitializer;
 import cloud.piranha.extension.soteria.SoteriaPreCDIInitializer;
 import cloud.piranha.extension.weld.WeldInitializer;
@@ -51,14 +59,21 @@ public class JakartaSecurityAllInitializer implements ServletContainerInitialize
 
         // Makes web.xml login-config available to Soteria
         new SoteriaPreCDIInitializer(),
-        
-        // Initializes the Servlet Security primitives, on which Jakarta Security is based
-        new ServletSecurityAllInitializer(),
+
+        // Configures the security constraints, authorization module
+        // and authorization filter that checks constraints before authentication
+        new AuthorizationPreInitializer(),
+
+        // Configures the authentication module and authentication filter
+        new AuthenticationInitializer(),
+
+        // Configures the authorization filter that checks constraints after authentication
+        new AuthorizationInitializer(),
 
         // Initializes CDI, on which Jakarta Security is also based
         new WeldInitializer(),
 
-        // Configures Soteria, which implements Jakarta Security and sets itself as a Servlet 
+        // Configures Soteria, which implements Jakarta Security and sets itself as a Servlet
         // authentication module
         new SoteriaInitializer(),
     };
@@ -72,6 +87,9 @@ public class JakartaSecurityAllInitializer implements ServletContainerInitialize
      */
     @Override
     public void onStartup(Set<Class<?>> classes, ServletContext servletContext) throws ServletException {
+        servletContext.setAttribute(AUTHZ_FACTORY_CLASS, DefaultPolicyConfigurationFactory.class);
+        servletContext.setAttribute(AUTHZ_POLICY_CLASS, DefaultPolicy.class);
+
         for (ServletContainerInitializer initializer : initializers) {
             initializer.onStartup(classes, servletContext);
         }
